pyproject.toml
[project]
name = "cede"
version = "0.0.0"
description = "CEDE lab tooling (lab.yaml validation, Raspberry Pi bootstrap helpers)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv]
package = false
